+++ lib/src/models/models.dart
@@ -3,3 +3,4 @@ export 'focus_event.dart';
 export 'app_info.dart';
 export 'focus_tracker_config.dart';
 export 'browser_tab_info.dart';
+export 'input_activity.dart';